backend/requirements.txt
aioredis==2.0.1
elasticsearch[async]==7.9.1
fastapi==0.94.1
gunicorn==20.1.0
pydantic==1.10.4
python-dotenv==0.21.1
uvicorn==0.20.0
uvloop==0.17.0
ormar[orjson]==0.12.1